+++ src/utils/context.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const MyContext = createContext();
