+++ src/utils/getRandomHexColor.jsx
@@ -1,0 +1,3 @@
+export function getRandomHexColor() {
+  return `#${Math.floor(Math.random() * 16777215).toString(16)}`;
+}
